+++ fiddles/react/fiddle-0002-FunctionalComponent/src/index.js
@@ -1,0 +1,22 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import './index.css';
+
+function Element(props) {
+  return <h1>{props.value}</h1>;
+}
+
+function App() {
+  return (
+    <div>
+      <Element value="1" />
+      <Element value="2" />
+      <Element value="3" />
+    </div>
+  );
+}
+
+ReactDOM.render(
+  <App />,
+  document.getElementById('root')
+);