+++ fiddles/react/fiddle-0004-AppComponent/src/App.js
@@ -1,0 +1,19 @@
+import React, { Component } from 'react';
+import './App.css';
+import './clock/Clock';
+import Clock from "./clock/Clock";
+
+
+class App extends Component {
+  render() {
+    return (
+      <div>
+        <Clock/>
+        <Clock/>
+        <Clock/>
+      </div>
+    );
+  }
+}
+
+export default App;