+++ fiddles/react/fiddle-0005-EventHandler/src/App.js
@@ -1,0 +1,82 @@
+import React, { Component } from 'react';
+import './toggle/Toggle';
+import Toggle from "./toggle/Toggle";
+
+class App extends Component {
+  render() {
+    return (
+      <div>
+       <div>
+         <Toggle/>
+         <Toggle/>
+         <Toggle/>
+         <Toggle/>
+         <Toggle/>
+         <Toggle/>
+         <Toggle/>
+         <Toggle/>
+         <Toggle/>
+       </div>
+        <div>
+          <Toggle/>
+          <Toggle/>
+          <Toggle/>
+          <Toggle/>
+          <Toggle/>
+          <Toggle/>
+          <Toggle/>
+          <Toggle/>
+          <Toggle/>
+        </div>
+        <div>
+          <Toggle/>
+          <Toggle/>
+          <Toggle/>
+          <Toggle/>
+          <Toggle/>
+          <Toggle/>
+          <Toggle/>
+          <Toggle/>
+          <Toggle/>
+        </div>
+        <div>
+          <Toggle/>
+          <Toggle/>
+          <Toggle/>
+          <Toggle/>
+          <Toggle/>
+          <Toggle/>
+          <Toggle/>
+          <Toggle/>
+          <Toggle/>
+        </div>
+        <div>
+          <Toggle/>
+          <Toggle/>
+          <Toggle/>
+          <Toggle/>
+          <Toggle/>
+          <Toggle/>
+          <Toggle/>
+          <Toggle/>
+          <Toggle/>
+        </div>
+        <div>
+          <Toggle/>
+          <Toggle/>
+          <Toggle/>
+          <Toggle/>
+          <Toggle/>
+          <Toggle/>
+          <Toggle/>
+          <Toggle/>
+          <Toggle/>
+        </div>
+      </div>
+
+
+    );
+  }
+}
+
+export default App;