+++ fiddles/react/fiddle-0006-ConditionalRendering/src/App.js
@@ -1,0 +1,15 @@
+import React, { Component } from 'react';
+import './App.css';
+import LoginControl from './login-control/LoginControl';
+
+class App extends Component {
+  render() {
+    return (
+      <div className="App">
+       <LoginControl/>
+      </div>
+    );
+  }
+}
+
+export default App;