+++ fiddles/react/fiddle-0007-HiddenRendering/src/page/Page.test.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import Page from './Page';
+
+it('renders without crashing', () => {
+  const div = document.createElement('div');
+  ReactDOM.render(<Page />, div);
+  ReactDOM.unmountComponentAtNode(div);
+});