+++ scripts/bin/react/_create.sh
@@ -120,8 +120,8 @@ function create() {
   echo "$0" | sed 's/\.\///g' | awk '{print toupper($0)}'
   echo "Bash version ${BASH_VERSION}..."
 
-  fiddleSubDir="../fiddles/angular2-cli/$1";
-  fiddleTemplateDir="../fiddles/angular2-cli/template";
+  fiddleSubDir="../fiddles/react/$1";
+  fiddleTemplateDir="../fiddles/react/template";
   bornOnDate=$(date +"%m-%d-%y";)
   echo ${bornOnDate};